rename CSV download modal file and hide downlaod icon on sites page
